+++ splib/Testing.py
@@ -9,45 +9,9 @@ from splib.topology.loader import *
 from splib.core.node_wrapper import *
 
 from splib.core.node_wrapper import ChildWrapper, ObjectWrapper
+from splib.helper import exportScene
 
 
-class displayNode():
-    def __init__(self,_level=0):
-        self.prefix = ""
-        for i in range(_level):
-            self.prefix += "| "
-
-    def addObject(self,type:str,**kwargs):
-        print(self.prefix + type + " with " + str(kwargs))
-
-    def addChild(self,name:str):
-        print(self.prefix + "-> Node : " + name)
-        return displayNode(len(self.prefix) + 1)
-
-class exportScene():
-    def __init__(self,name="rootNode"):
-        self.name = name
-
-    def addObject(self,type:str,**kwargs):
-        suffix = ""
-        for i in kwargs:
-            suffix += "," + str(i) + "=\"" + str(kwargs[i]) + "\""
-        print(self.name+".addObject(\"" + type + "\"" +  suffix + ")")
-
-    def addChild(self,name:str):
-        print(name + '=' + self.name+".addChild(\"" + name + "\")")
-        setattr(self,name,exportScene(name))
-        return getattr(self,name)
-
-    def __setattr__(self, key, value):
-        if(not(key == "name")):
-            print(self.__dict__["name"] + "." + key + " = " + str(value))
-            self.__dict__[key] = value
-        else:
-            self.__dict__[key] = value
-
-
-@PrefabSimulation
 def createScene(rootNode):
     rootNode.dt = 0.03
     rootNode.gravity = [0,-9.81,0]
